Refactor balances
